see #2: move non-type template param auto to ch02
