+++ code/2/2.14.default.template.param.cpp
@@ -1,0 +1,18 @@
+//
+// 2.4.default.template.param.cpp
+// chapter 2 language usability
+// modern cpp tutorial
+//
+// created by changkun at changkun.de
+//
+
+#include <iostream>
+
+template<typename T = int, typename U = int>
+auto add(T x, U y) -> decltype(x+y) {
+    return x+y;
+}
+
+int main() {
+    std::cout << add(1, 2) << std::endl;
+}
